+++ kryptokrona-http/src/main/kotlin/org/kryptokrona/sdk/http/model/response/block/Blocks.kt
@@ -36,7 +36,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Blocks(
     @SerialName("response.blocks") val blocks: List<String>, // TODO: should probably be a list of Block
-    @SerialName("response.current_height") val currentHeight: Int,
-    @SerialName("response.start_height") val startHeight: Int,
+    @SerialName("response.current_height") val currentHeight: Long,
+    @SerialName("response.start_height") val startHeight: Long,
     @SerialName("response.status") val status: String
 )
